+++ VillageProphecy/VillageProphecy/EnemySkill.h
@@ -7,7 +7,7 @@ using namespace std;
 struct SkillEffect
 {
 	SkillConsumeableStats consumes;
-	float amountConsumed;
+	int amountConsumed;
 	string damageDescription;
 };
 
@@ -17,7 +17,7 @@ class EnemySkill
 public:
 	virtual bool isDebuff() = 0;
 	virtual string getSkillName() = 0;
-	virtual float getSkillDamage() = 0;
+	virtual int getSkillDamage() = 0;
 	virtual bool CanCast() = 0;
 	virtual SkillEffect getEffect() = 0;
 	virtual int roundsToLast() = 0;
